+++ src/shared/preferences.js
@@ -9,7 +9,11 @@ import { VALID_THEMES } from './utils.js';
 /**
  * Valid position options for notes
  */
-export const VALID_POSITIONS = ['top-left', 'top-right', 'bottom-left', 'bottom-right'];
+export const VALID_POSITIONS = [
+  'top-left', 'top-center', 'top-right',
+  'center-left', 'center-right',
+  'bottom-left', 'bottom-center', 'bottom-right'
+];
 
 /**
  * Valid note width options (in pixels)